install python-is-python3
`python-is-python3` just creates a symlink from /bin/python to the python3 binary on the system. There are too many 3rd party scripts/packages that reference the universal `python` binary so this keeps the fallout to a minimum. We do this because we are explicitly removing `python2` which comes by default from upstream Debian (in the package aptly named `python-is-python2`)